+++ src/components/Datepicker/Datepicker.js
@@ -382,7 +382,7 @@ export class Datepicker extends React.Component {
   }
 
   render() {
-    const { required } = this.props;
+    const { className, disabled, required } = this.props;
     const { inputValue, isValid, open, viewedDate } = this.state;
     const viewedMonthName = moment(viewedDate).format('MMMM');
 
@@ -392,8 +392,9 @@ export class Datepicker extends React.Component {
     const placeholder = moment().localeData().longDateFormat(placeholderDateFormat);
 
     return (
-      <div style={{ position: 'relative' }}>
+      <div className={className} style={{ position: 'relative' }}>
         <Input
+          disabled={disabled}
           id="date-input"
           label={inputFieldLabel}
           placeholder={placeholder}
@@ -443,7 +444,9 @@ export class Datepicker extends React.Component {
 }
 
 Datepicker.defaultProps = {
+  className: null,
   date: undefined,
+  disabled: false,
   initialDate: null,
   locale: 'en',
   open: false,
@@ -455,9 +458,17 @@ Datepicker.defaultProps = {
 
 Datepicker.propTypes = {
   /**
+   * Additional class name
+   */
+  className: PropTypes.string,
+  /**
    * Date in ISO 8601 format (controlled component)
    */
   date: PropTypes.string,
+  /**
+   * Disabled state
+   */
+  disabled: PropTypes.bool,
   /**
    * Optional initial date in ISO 8601 format
    */