+++ CHANGELOG.md
@@ -21,6 +21,24 @@
 * Breaking: `Box` theme usage
 * Breaking: Disallow `IconSVG` to have a background
 * Remove circle boolean, LDS sets it automatically for `action` icons
+=======
+## 2.1.1
+
+* Enhance Lookups with optional labels, add required and error state
+
+## 2.1.0
+
+* Additional display props for Datepicker (`className` and `disabled`)
+* Enable tooltips for menu
+* Add custom renderer callback for `Lookup` pills
+* Bugfix: rendering of mulitple columns in `Lookup` in table mode
+
+## 2.0.3
+
+* Add option for disabling error message display
+* Add read only Inputs
+* Bugfix: Pill without close Button
+* Bugfix: Fix IconSVG background
 
 ## 2.0.2